Shift Desktop Icons - 0 CP
Yes, I added the ability to customize the Desktop Icons in the Shifter.
This is very basic, and I tried my best to allow older skins to be
converted to enable these abilities. Don't worry though, the title was
just a creative title and no upgrades were added to the Shiftorium. I
also added the ability to delete files from the desktop.