Experimental Features...
LOADS of experimental stuff with Catalyst, Skinning and more!

Why not create a Sounds directory in your Skins\Loaded directory, and
stick a few .wav files in? Preferrably an infobox.wav, startup.wav and
shutdown.wav. Soon, there will be ways of adding these files in-game
through the Shifter, don't worry. Also, a Catalyst Story can be found in
Helper.vb in the form of a string array.