Huge bug fixes
This commit fixes some major bugs, atleast I hope so, that caused
ShiftOS to be unplayable. The bugs I'm talking about are the inability
to find the xulrunner, MichaelsMovableControls.dll, and
SkyBound.Gecko.dll files.

I also fixed a minor bug where when .smf files are loaded in the File
Skimmer or Desktop, the modloader would read the code rather than
opening the file.